backend_python/requirements.txt
aiofiles==24.1.0
anyio==4.8.0
asgiref==3.8.1
async-property==0.2.2
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coverage==6.3.2
cryptography==44.0.2
debugpy==1.6.0
deprecation==2.1.0
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
Markdown==3.7
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
python-keycloak==5.3.1
PyYAML==6.0.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.0
urllib3==2.3.0
